created user scripts, added redirect to user from searchbar
