+++ docs/js/index.js
@@ -195,68 +195,6 @@ let leaderboard = (function(){
    }
 })();
 
-//TODO: move all this to a custom page
-let userAccount = (function(){
-   let ids = {
-      searchBar: {
-         desktop: 'investor-username',
-         mobile: 'investor-username-mobile'         
-      },
-      searchButton: {
-         desktop: 'investor-username-search',
-         mobile: undefined
-      }
-   }
-   function show(device) {
-      let username = document.getElementById(ids.searchBar[device]).value;
-      if(username.length > 0){
-         let domPopup = document.getElementById('investor-info'); 
-         M.Modal.init(domPopup);
-         M.Modal.getInstance(domPopup).open()
-         jsonApi.get('/investor/'+username).then(function(data) {
-         document.getElementById('investor-account-data').innerHTML = `
-              <h5>${username}'s profile</h5>
-              <p><a target="_blank" href="https://reddit.com/u/${username}">visit reddit profile</a></p>
-              <table>
-                 <tr><th>Balance</th><td>${data.balance}</td></tr>
-                 <tr><th>Gone broke</th><td>${data.broke} times</td></tr>
-                 <tr><th>Investments</th><td>${data.completed}</td></tr>
-              </table>`;
-         })
-         .catch(function(er){
-            if (er.status === 404){
-               document.getElementById('investor-account-data').innerHTML = `<h5>There is no investor with that username!</h5>`
-            }
-         });
-      }
-   }
-   function init(){
-      //add ENTER key listener
-      let searchEl = document.getElementById(ids.searchBar['desktop']);
-      searchEl.addEventListener('keypress', e=> {if((e.which || e.keyCode) === 13)show('desktop')} );     
-      searchEl = document.getElementById(ids.searchBar['mobile']);
-      searchEl.addEventListener('keypress', e=> {if((e.which || e.keyCode) === 13)show('mobile')} );
-      //add SEARCH button click listener
-      let searchButton = document.getElementById(ids.searchButton['desktop']);
-      searchButton.addEventListener('click', e=> show('desktop') );  
-      // check if url contains ?account=
-      let url = new URL(window.location.href);
-      let username = url.searchParams.get("account");
-      const reg = /^[a-zA-Z0-9\-\_]+$/;
-      if (reg.test(username)) {
-         document.getElementById('investor-username').value = username;
-         show('desktop');
-         history.pushState(null, '', window.location.href.split('?')[0]);
-      }
-
-   }
-   return {
-      init: init,
-      show: show
-   }
-})();
-
-
 let investmentsCalculator = (function() {
    let startEl, endEl, amountEl;
    let ids = {
@@ -379,7 +317,6 @@ let updater = (function(){
       //init local modules
       overview.init()
       overviewChart.init()
-      userAccount.init()
       updater.init()
       investmentsCalculator.init()
