+++ docs/user.html
@@ -18,8 +18,8 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/Chart.js/2.4.0/Chart.min.js"></script>
     <script src="js/libs/countup.min.js"></script>
     <!-- page -->
-    <link type="text/css" rel="stylesheet" href="css/index.css"/>
-    <script type="module" src="js/index.js"></script>
+    <link type="text/css" rel="stylesheet" href="css/user.css"/>
+    <script type="module" src="js/user.js"></script>
 
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
 </head>
@@ -63,16 +63,19 @@
         <!-- <li><a href="foo.html">brokers</a></li> -->
         <!-- <li><a href="bar.html">commands</a></li> -->
         <li><a href="https://github.com/MemeInvestor/memeinvestor_bot/issues" target="_blank">bugs</a></li>
-        <li><a href="https://github.com/MemeInvestor/memeinvestor_bot/"><img class="icon-32" src="resources/GitHub-Mark-Light-32px.png"
-                                               target="_blank"></a></li>
+        <li><a href="https://github.com/MemeInvestor/memeinvestor_bot/">
+        <img class="icon-32" src="resources/GitHub-Mark-Light-32px.png" target="_blank"></a></li>
     </ul>
     
     <div id="announcements-modal" class="modal ">
     </div>
 
-  <!-- PUT PAGE CODE HERE -->
-  
-  
+    <!-- PUT PAGE CODE HERE --> 
     
+    
+    <div class="container" >
+      <div id="investor-account-data" class="table-wrap"></div>
+    </div>
+  
   </body>  
 </html>
